Refactoring: fix circular imports
